partition balancer

+++ src/main/java/com/sanketika/models/PipelineMeta.java
@@ -2,12 +2,12 @@ package com.sanketika.models;
 
 public class PipelineMeta {
     private long arrival_tstamp;
-    private int preprocessor_time_ms;
+    private long preprocessor_time_ms;
 
     public PipelineMeta() {
     }
 
-    public PipelineMeta(long arrival_tstamp, int preprocessor_time_ms) {
+    public PipelineMeta(long arrival_tstamp, long preprocessor_time_ms) {
         this.arrival_tstamp = arrival_tstamp;
         this.preprocessor_time_ms = preprocessor_time_ms;
     }
@@ -20,11 +20,11 @@ public class PipelineMeta {
         this.arrival_tstamp = arrival_tstamp;
     }
 
-    public int getPreprocessor_time_ms() {
+    public long getPreprocessor_time_ms() {
         return preprocessor_time_ms;
     }
 
-    public void setPreprocessor_time_ms(int preprocessor_time_ms) {
+    public void setPreprocessor_time_ms(long preprocessor_time_ms) {
         this.preprocessor_time_ms = preprocessor_time_ms;
     }
 }
